requirements.txt
fastapi==0.111.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0